backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Environment & config
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Document parsing
pypdf==4.3.1
python-multipart==0.0.12

# NLP / chunking
langchain-text-splitters==0.3.0

# Embeddings (runs locally, no API key needed)
sentence-transformers==3.1.1

# Vector database
chromadb==0.5.20

# LLM
anthropic==0.34.2

# Utilities
httpx==0.27.2
tenacity==9.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
